Pabeigta OIzveide klase

+++ src/uzdevums5/Trijsturis.java
@@ -5,7 +5,7 @@ public class Trijsturis {
 	
 	public Centrs punkts;
 	
-	public Trijsturis(Centrs p, double m1, double m2, double m3) {
+	public Trijsturis(Centrs p, int m1, int m2, int m3) {
 		punkts = p;
 		mala1 = m1;
 		mala2 = m2;
